requirements.txt
Django==4.2.29
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
openai==1.14.3
openpyxl==3.1.2
pandas==2.2.1
matplotlib==3.8.3
Pillow==10.2.0
python-dotenv==1.0.1